# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is the admin API for the Urchin app.",
        "title": "GoCMS Admin API",
        "contact": {
            "name": "Ricardo",
            "email": "ricardobenthem@gmail.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
...
        "/posts": {
            "get": {
                "description": "Retrieves a paginated list of posts.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "posts"
                ],
                "summary": "Get a list of posts",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Offset for pagination",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Limit for pagination. 0 means no limit.",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/admin_app.GetPostsResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid offset or limit parameter",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates an existing post with new data.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "posts"
                ],
                "summary": "Update an existing post",
                "parameters": [
                    {
                        "description": "Post data to update",
                        "name": "post",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admin_app.ChangePostRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/admin_app.PostIdResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body or could not change post",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Adds a new post to the database.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "posts"
                ],
                "summary": "Add a new post",
                "parameters": [
                    {
                        "description": "Post to add",
                        "name": "post",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admin_app.AddPostRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/admin_app.PostIdResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body or missing data",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a post by its ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "posts"
                ],
                "summary": "Delete a post",
                "parameters": [
                    {
                        "description": "Post ID to delete",
                        "name": "id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admin_app.DeletePostRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/admin_app.PostIdResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid ID provided",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Post not found",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/posts/{id}": {
            "get": {
                "description": "Retrieves a single post by its ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "posts"
                ],
                "summary": "Get a single post",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Post ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/admin_app.GetPostResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid post ID",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Post not found",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a page by its Link.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pages"
                ],
                "summary": "Delete a page",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Page Link",
                        "name": "link",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/admin_app.DeletePageRequest"
                        }
                    },
                    "400": {
                        "description": "Invalid link provided",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Page not found",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "admin_app.AddPostRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "description": "Content of the post\nin: body",
                    "type": "string"
                },
                "excerpt": {
                    "description": "Excerpt of the post\nin: body",
                    "type": "string"
                },
                "title": {
                    "description": "Title of the post\nin: body\nrequired: true",
                    "type": "string"
                }
            }
        },
...
        "admin_app.ChangePostRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "description": "Content of the post\nin: body",
                    "type": "string"
                },
                "excerpt": {
                    "description": "Excerpt of the post\nin: body",
                    "type": "string"
                },
                "id": {
                    "description": "ID of the post\nin: body\nrequired: true",
                    "type": "integer"
                },
                "title": {
                    "description": "Title of the post\nin: body",
                    "type": "string"
                }
            }
        },
        "admin_app.DeletePageRequest": {
            "type": "object",
            "properties": {
                "link": {
                    "description": "Link of the page to delete\nin: body\nrequired: true",
                    "type": "string"
                }
            }
        },
        "admin_app.DeletePostRequest": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
...
        "admin_app.GetPostResponse": {
            "type": "object",
            "properties": {
                "content": {
                    "description": "Content of the post",
                    "type": "string"
                },
                "excerpt": {
                    "description": "Excerpt of the post",
                    "type": "string"
                },
                "id": {
                    "description": "ID of the post",
                    "type": "integer"
                },
                "title": {
                    "description": "Title of the post",
                    "type": "string"
                }
            }
        },
        "admin_app.GetPostsResponse": {
            "type": "object",
            "properties": {
                "posts": {
                    "description": "List of posts",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/common.Post"
                    }
                }
            }
        },
...
        "admin_app.PostIdResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "ID of the post",
                    "type": "integer"
                }
            }
        },
...
        "common.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "common.Post": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "excerpt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        }
    }
}